// frontend-admin/src/api/generated/model/adminOnlinePaymentDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */

export interface AdminOnlinePaymentDto {
  amount?: number;
  /** @nullable */
  completedAtUtc?: string | null;
  createdAtUtc?: string;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  id?: string;
  isSynthetic?: boolean;
  /** @nullable */
  lastWebhookEvent?: string | null;
  /** @nullable */
  onlineOrderId?: string | null;
  /** @nullable */
  orderNumber?: string | null;
  /** @nullable */
  paymentIntentId?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  status?: string | null;
  tenantId?: string;
  /** @nullable */
  tenantName?: string | null;
  /** @nullable */
  tenantSlug?: string | null;
}