// frontend-admin/src/api/generated/model/adminOnlinePaymentListResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */
import type { AdminOnlinePaymentDto } from './adminOnlinePaymentDto';

export interface AdminOnlinePaymentListResponse {
  /** @nullable */
  items?: AdminOnlinePaymentDto[] | null;
  pageNumber?: number;
  pageSize?: number;
  totalCount?: number;
}