// frontend-admin/src/api/generated/model/adminOnlinePaymentTestResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */
import type { OnlinePaymentDto } from './onlinePaymentDto';
import type { AdminOnlinePaymentDto } from './adminOnlinePaymentDto';

export interface AdminOnlinePaymentTestResponse {
  /** @nullable */
  code?: string | null;
  /** @nullable */
  error?: string | null;
  posPayment?: OnlinePaymentDto;
  succeeded?: boolean;
  transaction?: AdminOnlinePaymentDto;
}